+++ app/controllers/all/users/projects_controller.rb
@@ -1,0 +1,9 @@
+class All::Users::ProjectsController < ApplicationController
+  after_action :verify_authorized
+  rescue_from ActiveRecord::RecordNotFound, with: :not_found_error
+
+  def index
+    authorize Project, :index?
+    @users = ByUserProjectFetcherService.new.call
+  end
+end