+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added a view, by region, that shows the region with in-progress projects
 - By region view links to the projects in that region
 - Add a rake task to update the `team` attribute on all existing projects
+- Added a new view that list users who are assigned to projects, showing the
+  number of conversions
 
 ### Changed